move test utility classes to utils package for better organization
