+++ application/src/test/java/org/mandarin/booking/adapter/security/CustomAuthenticationProviderTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
-import org.mandarin.booking.IntegrationTest;
 import org.mandarin.booking.adapter.CustomMemberAuthenticationToken;
+import org.mandarin.booking.utils.IntegrationTest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
