+++ application/src/test/java/org/mandarin/booking/utils/TestResult.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking;
+package org.mandarin.booking.utils;
 
 import static org.assertj.core.api.Assertions.fail;
