+++ application/src/test/java/org/mandarin/booking/webapi/auth/reissue/POST_specs.java
@@ -1,15 +1,15 @@
 package org.mandarin.booking.webapi.auth.reissue;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mandarin.booking.JwtTestUtils.assertJwtFormat;
-import static org.mandarin.booking.JwtTestUtils.getExpiration;
 import static org.mandarin.booking.MemberAuthority.USER;
 import static org.mandarin.booking.adapter.ApiStatus.BAD_REQUEST;
 import static org.mandarin.booking.adapter.ApiStatus.SUCCESS;
 import static org.mandarin.booking.adapter.ApiStatus.UNAUTHORIZED;
-import static org.mandarin.booking.fixture.MemberFixture.NicknameGenerator.generateNickName;
-import static org.mandarin.booking.fixture.MemberFixture.PasswordGenerator.generatePassword;
-import static org.mandarin.booking.fixture.MemberFixture.UserIdGenerator.generateUserId;
+import static org.mandarin.booking.utils.JwtTestUtils.assertJwtFormat;
+import static org.mandarin.booking.utils.JwtTestUtils.getExpiration;
+import static org.mandarin.booking.utils.MemberFixture.NicknameGenerator.generateNickName;
+import static org.mandarin.booking.utils.MemberFixture.PasswordGenerator.generatePassword;
+import static org.mandarin.booking.utils.MemberFixture.UserIdGenerator.generateUserId;
 
 import io.jsonwebtoken.security.Keys;
 import java.util.Date;
@@ -18,11 +18,11 @@ import javax.crypto.SecretKey;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.mandarin.booking.IntegrationTest;
-import org.mandarin.booking.IntegrationTestUtils;
 import org.mandarin.booking.TokenHolder;
 import org.mandarin.booking.adapter.TokenUtils;
 import org.mandarin.booking.domain.member.ReissueRequest;
+import org.mandarin.booking.utils.IntegrationTest;
+import org.mandarin.booking.utils.IntegrationTestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.TestPropertySource;
