+++ application/src/test/java/org/mandarin/booking/webapi/not_found/GET_specs.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mandarin.booking.adapter.ApiStatus.NOT_FOUND;
 
 import org.junit.jupiter.api.Test;
-import org.mandarin.booking.IntegrationTest;
-import org.mandarin.booking.IntegrationTestUtils;
-import org.mandarin.booking.NoRestDocs;
+import org.mandarin.booking.utils.IntegrationTest;
+import org.mandarin.booking.utils.IntegrationTestUtils;
+import org.mandarin.booking.utils.NoRestDocs;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @IntegrationTest
